requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
ipython==8.12.0
jedi==0.18.2
jsonschema==4.17.3
loguru==0.7.0
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
openapi==1.1.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pinecone-client==2.2.1
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
stack-data==0.6.2
tiktoken==0.3.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.8.2
